+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 Use the following format for additions: ` - VERSION: [feature/patch (if applicable)] Short description of change. Links to relevant issues/PRs.`
 
+- 1.1.21:
+  - Initial refactoring of click test using nightmare and mocha
+  - **Dropping support for node 4.x and 5.x!, 6.x and later is now supported.**
 - 1.1.20: Hide credentials in remote urls at home repo list
 - 1.1.19: Ask before deleting a stash
 - 1.1.18: Fix checking out remote refs (again)